waipio: dts: remove wsa link from audo route
move this audio route to machine driver as the wsa device is not
unknown. For example, the right wsa883x chip is on the sub board.
Maybe it's not connected to main board. we need depend probe result
to dynamically add audio route.

Change-Id: Ida9b02534f06721e169ca377a6d392493cbd9d5d
Signed-off-by: huangys <huangys@motorola.com>
Reviewed-on: https://gerrit.mot.com/2180980
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Haichao Li <liuhc3@motorola.com>
Reviewed-by: Guobin Zhang <zhanggb@motorola.com>